run test in node and bun
